requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.2
django-cors-headers==4.7.0
django-filter==22.1
django-graphql-jwt==0.4.0
djangorestframework==3.16.0
graphene==3.4.3
graphene-django==3.2.3
graphene-django-extras==1.0.0
graphene-file-upload==1.3.0
graphql-core==3.2.6
graphql-relay==3.2.0
gunicorn==23.0.0
iniconfig==2.1.0
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
promise==2.3
psycopg2-binary==2.9.10
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.4.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.13.2
tzdata==2025.2
whitenoise==6.9.0
